Moves _skip_pron to extract/core.py
