+++ tests/test_extract.py
@@ -2,12 +2,12 @@ import pytest
 import requests
 
 from wikipron.extract import EXTRACTION_FUNCTIONS
+from wikipron.extract.core import _skip_pron
 from wikipron.extract.default import extract_word_pron_default
 
 
 @pytest.mark.parametrize(
-    "func",
-    tuple(EXTRACTION_FUNCTIONS.values()) + (extract_word_pron_default,),
+    "func", tuple(EXTRACTION_FUNCTIONS.values()) + (extract_word_pron_default,)
 )
 def test_extraction_functions_have_the_same_signature(func):
     expected_annotations = {
@@ -20,3 +20,15 @@ def test_extraction_functions_have_the_same_signature(func):
     assert expected_annotations == actual_annotations, (
         f"{func.__qualname__} does not have the expected function signature.",
     )
+
+
+@pytest.mark.parametrize(
+    "pron, expected",
+    [
+        ("əbzɝvɚ", True),
+        # GH-105: Dashed prons are skipped.
+        ("ɑb-", False),
+    ],
+)
+def test__skip_pron(pron, expected):
+    assert _skip_pron(pron) == expected
